BE-build/requirements.txt
setuptools==65.5.1
sourcedefender
wheel==0.38.1
aiofiles==0.8.0
aioredis==1.3.1
apispec==5.1.1
apispec-webframeworks==0.5.2
APScheduler==3.6.3
asgiref==3.3.4
async-timeout==4.0.2
autobahn==22.12.1
Automat==20.2.0
beautifulsoup4==4.9.1
boltons==21.0.0
cachetools==4.1.1
certifi==2023.7.22
cffi==1.14.4
channels==3.0.3
channels-redis==3.3.0
chardet==3.0.4
charset-normalizer==2.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
daphne==3.0.2
deepdiff==5.3.0
Django==3.2.20
django-apscheduler==0.5.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.5.0
django-sslserver==0.22
djangorestframework==3.11.2
docopt==0.6.2
drf-yasg==1.20.0
environs==9.5.0
et-xmlfile==1.0.1
feedparser==6.0.8
gcloud==0.18.3
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==2.0.1
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.19.0
httptools==0.3.0
hyperlink==21.0.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0
jsonschema==3.2.0
jws==0.1.3
lark-parser==0.12.0
logger==1.4
MarkupSafe==2.1.1
marshmallow==3.7.1
msgpack==1.0.4
multidict==5.2.0
numpy==1.22.2
oauth2==1.9.0.post1
oauth2client==4.1.2
openapi-codec==1.3.2
openpyxl==3.0.5
ordered-set==4.0.2
packaging==21.3
pandas==1.2.4
pdfkit==0.6.1
pendulum==2.1.2
pkcs7==0.1.2
protobuf==3.18.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.4.0
pymongo==3.11.4
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-jose==3.3.0
pytzdata==2020.1
redis==4.4.4
requests==2.31.0
requests-toolbelt==0.9.1
required==0.4.0
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==21.1.0
sgmllib3k==1.0.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.4
swagger-spec-validator==2.7.3
TgCrypto==1.2.3
Twisted==23.8.0
txaio==21.2.1
typing-extensions==4.0.1
tzlocal==4.0
uritemplate==3.0.1
urllib3==1.26.5
websockets==9.1
xlrd==2.0.1
zope.event==4.5.0
zope.interface==5.2.0
google-auth-oauthlib==0.4.6
xmltodict==0.12.0
mailchimp_marketing==3.0.74
mail-parser==3.15.0
url-normalize==1.4.3
colorlog==6.6.0
simpleeval==0.9.12
pytz==2022.1
celery[redis]==5.2.6
django-celery-results==2.4.0
celery-singleton==0.3.1
imapclient==2.2.0
markdown==3.3.7
django-redis==5.2.0
mongomock==4.0.0
ldap3==2.9.1
Faker==13.14.0
o365==2.0.19
python-json-logger==2.0.4
django-request-logging==0.7.5
django-cache-memoize==0.1.10
django-health-check==3.17.0
psutil==5.9.2
django-ipware==4.0.2
geoip2==4.6.0
ua-parser==0.16.1
user-agents==2.2.0
django-user-agents==0.4.0
django-celery-beat==2.4.0
celery-redbeat==2.1.0
gevent==22.10.2
elasticsearch_dsl==7.4.0
django-ratelimit==4.0.0
aioxmpp==0.13.3
weasyprint==59.0
exchangelib==5.0.3
ipython==8.12.2
XlsxWriter==3.1.2
pypdf==3.13.0
Unidecode==1.3.6